Builder/Adapter

+++ Seminar5/FactoryMetod.PB3.RESTAURANT/src/fabrici/TipMancareFactory.java
@@ -3,5 +3,5 @@ package fabrici;
 import clase.FelMancare;
 
 public interface TipMancareFactory {
-    FelMancare preparareFelMancare(TipFelMancare tip, int pret, int nrCalorii);
+    FelMancare preparareFelMancare(TipFelMancare tip, float pret, int nrCalorii);
 }
